Library files reorganisation rational
I have reorganised the files for the library for two main reasons
1) Make it simpler to allow none std::strings to be integral
2) Remove the string header from the DSL headers


1) Make it simpler to allow none std::strings to be integral
OOLUA_STD_STRING_IS_INTEGRAL
When this macro is defined to be one, as has always been the default, it allows a std::string to pulled from the stack by the DSL without the user defining a proxy for it. I have attempt a couple of things in the past to make it easier to have for example wxString treated the same way; yet this involved creating all the traits for the type. Previous to this comment there was a macro in param_traits.h for this purpose, oolua_integral_trait, which you could then use to provide the following:

namespace OOLUA
{
        oolua_integral_trait(std::string, std::string, in_p, 1, 0)
        oolua_integral_trait(std::string&, std::string, in_p, 1 , 0)
        oolua_integral_trait(std::string const, std::string, in_p, 1, 0)
        oolua_integral_trait(std::string const&, std::string, in_p, 1, 0)

        template<>struct out_p<std::string>;//disable
        oolua_integral_trait(std::string &, std::string, out_p, 0, 1)
        oolua_integral_trait(std::string &, std::string, in_out_p, 1, 1)
        namespace INTERNAL
        {
                oolua_integral_trait(std::string, std::string, function_return, 0, 1)
                oolua_integral_trait(std::string&, std::string, function_return, 0, 1)
                oolua_integral_trait(std::string const, std::string, function_return, 0, 1)
                oolua_integral_trait(std::string const&, std::string, function_return, 0, 1)
        } // namespace INTERNAL
} // namespace OOLUA

Whilst this made it somewhat simpler and provided traits to treat the type as an integral for the DSL, it was never to my liking. In addition you still needed to provide OOLUA::push and OOLUA::LUA_CALLED::pull, LUA_CALLED::pull recently became LUA_CALLED::get, for the string type. There was also another way, which I have personally used, where you still provide the traits somewhat like above yet the type pulled from the stack is still std::string, I would then provide a OOLUA::INTERNAL::Convertor from std::string to wxString. What I really wanted was just one location such as a header and it's source file to make these user modifications.

2) Remove the string header from the DSL headers
If we take the following dummy class for example
struct foo{};

This could be proxied using the following
#include "oolua.h"
OOLUA_PROXY(foo)
OOLUA_PROXY_END

If the header were rewrote to the following, then I am sure some people would be up in arms and I would be in agreement with them

#include "oolua.h"
#include <string>
OOLUA_PROXY(foo)
OOLUA_PROXY_END

Why have you included the string header here? It is not needed, remove it!
Well this is what was happening anyway. oolua.h, lua_table.h, param_traits.h, also if you are using exceptions, pull in the string header. There was something I could do, which is not to include anything in the DSL headers which in turn pulls in string.


So the library has been re-arranged, with string types and there operations in oolua_string.h, a DSL header is provided for defining proxies and a DSL export header for proxy function exports.

Hang on a minute I hear you cry, you have just moved the problem as it is still going to pull it in. Well actually that is not true. As it now stands, no part of the DSL headers pulls in the string header any more and you do not need to provide traits to use std::string as an integral.


The oolua_string header provides forward declarations of three functions for strings, which are all in the OOLUA::STRING namespace. STRING::push which is used for OOLUA::push calls, STRING::pull used for OOLUA::pull calls
and STRING::get which is used by the DSL when proxying function calls. It also provides member function checking to see if the type is a string type. Simply to be an integral string type, it must not have a proxy type implemented for the type and must provide a specific member function. The default for this member checking is c_str returning char const*, although there are other macro examples you can use.

--HG--
branch : OOLua-2